backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
scrapy==2.12.0
apscheduler==3.10.4
python-dotenv==1.0.1
supabase==2.9.1
beautifulsoup4==4.12.3
webdriver-manager==4.0.2
selenium==4.25.0
pydantic==2.9.2